requirements.txt
Django==2.2.1
django-model-utils==3.1.2
djangorestframework==3.8.2
pytz==2019.1
sqlparse==0.3.0
psycopg2-binary==2.8.4
Faker==4.1.1
factory-boy==2.12.0